Missing some ACL cleanup for write, use 1 tested function in both files
